+++ changelog.md
@@ -1,4 +1,7 @@
 **Changes:**
-- Improved tags
-- Added compatibility with [Serene Seasons](https://www.curseforge.com/minecraft/mc-mods/serene-seasons)
+- Changed tags in recipes
+    - `#c:foods/vegetables/tomato` to `#c:crops/tomato`
+    - `#c:foods/vegetables/lettuce` to `#c:crops/lettuce`
+- Added NeoForge update checker url
+- Added new wiki link to the Mod Menu metadata
 - Updated dependencies